backend_your_movies/requirements.txt
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-redis==5.2.0
django-rq==2.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
rq==1.13.0
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
six==1.16.0
sqlparse==0.4.3
times==0.7
tzdata==2023.3
